% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api2.R
\name{extr}
\alias{extr}
\title{One-dimensional Bayesian global minimization (EXTR)}
\usage{
extr(
  bp,
  ep,
  evaluations,
  model_evals = 6L,
  acc_y = 1e-04,
  acc_x = 1e-04,
  objective = "furasn",
  trace = TRUE
)
}
\arguments{
\item{bp, ep}{Interval endpoints.}

\item{evaluations}{Maximum evaluations (Fortran limit: 500).}

\item{model_evals}{Evaluations used to estimate the Wiener model
parameters (>= 6).}

\item{acc_y, acc_x}{Requested accuracy of the optimum value and location.}

\item{objective}{An R function taking a numeric vector and returning a
scalar, or the name of a compiled built-in objective
(see \code{\link[=builtin_objectives]{builtin_objectives()}}).}

\item{trace}{If \code{TRUE}, return all sampled points and values.}
}
\value{
An object of class \code{globalopt_result}: a list with elements
\code{best_x}, \code{best_f}, \code{evals}, \code{points} (matrix of evaluated points, or
\code{NULL} when \code{trace = FALSE}), \code{values}, \code{method}, \code{backend},
\code{best_iter}.
}
\description{
Global minimization of a one-dimensional function on an interval using a
Wiener-process model (routine \code{EXTR}).
}
\examples{
r <- extr(-1, 1, 50, objective = function(x) (x[1] - 0.3)^2)
r$best_x
}
