\name{globalopt-package}
\alias{globalopt-package}
\alias{furasn}
\alias{lp_tau_point}
\alias{mig2}
\alias{bayes1}
\alias{lpmin}
\alias{glopt}
\alias{unt}
\alias{exkor}
\alias{extr}
\alias{mivar4}
\alias{flexi}
\alias{reqp}
\alias{lbayes}
\alias{anal1}
\alias{anal2}
\alias{local_minimize}
\alias{local_minimize_optim}
\alias{local_minimize_optimx}
\alias{local_minimize_nloptr}
\alias{local_minimize_minqa}
\alias{benchmark_functions}
\alias{benchmark_optimizers}
\alias{benchmark_deoptim}
\alias{run_benchmarks}
\alias{write_benchmark_csv}
\alias{write_benchmark_markdown}
\title{globalopt: global optimization routines and benchmarks}
\description{
Standalone R implementations of GlobalMinimum-inspired optimization routines,
local minimizer adapters, and benchmark/reporting helpers.
}
\details{
The package includes global search routines (for example \code{mig2}, \code{bayes1},
\code{lpmin}, \code{glopt}, \code{lbayes}), analysis helpers (\code{anal1}, \code{anal2}),
and benchmark utilities (\code{run_benchmarks}, report writers).
}
\seealso{
\code{\link{run_benchmarks}}, \code{\link{benchmark_functions}},
\code{\link{benchmark_optimizers}}
}
\keyword{package}
