% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectives.R
\name{objectives}
\alias{objectives}
\alias{furasn}
\alias{fush5}
\alias{fush7}
\alias{fush10}
\alias{fuhar3}
\alias{fuhar6}
\alias{fubran}
\alias{fugold}
\title{Built-in test objectives (R implementations)}
\usage{
furasn(x)

fush5(x)

fush7(x)

fush10(x)

fuhar3(x)

fuhar6(x)

fubran(x)

fugold(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Objective value (numeric scalar).
}
\description{
R implementations of the test objectives shipped with the original
GlobalMinimum library: \code{furasn} (a rastrigin-like separable function,
global minimum -2 at the origin), the Shekel family \code{fush5}/\code{fush7}/
\code{fush10} (4-D), the Hartmann family \code{fuhar3} (3-D) and \code{fuhar6} (6-D),
\code{fubran} (Branin, 2-D) and \code{fugold} (Goldstein-Price, 2-D).
Identical compiled versions are available by passing the function name
as the \code{objective} argument of the optimizers (see
\code{\link[=builtin_objectives]{builtin_objectives()}}).
}
