% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\docType{package}
\name{globalopt-package}
\alias{globalopt}
\alias{globalopt-package}
\title{globalopt: Bayesian Global Optimization Methods from the 'GlobalMinimum' Library}
\description{
Bindings to the 'GlobalMinimum' Fortran library of Bayesian and other global optimization methods by Jonas Mockus (1989), including the one-step Bayesian method ('BAYES1'), Monte-Carlo baselines ('MIG1', 'MIG2'), Bayesian search with local refinement ('LBAYES'), clustering ('GLOPT'), uniform deterministic search ('UNT'), one-dimensional and coordinate-wise Wiener-model search ('EXTR', 'EXKOR'), LP-tau deterministic search ('LPMIN'), a variable-metric local method ('MIVAR4'), a flexible-tolerance constrained simplex ('FLEXI'), recursive quadratic programming ('REQP'), and variable-screening procedures ('ANAL1', 'ANAL2'). The original Fortran code is compiled into the package; objectives can be supplied as R functions or chosen from the built-in compiled test objectives. A pure-R reference implementation of the core methods is included for comparison.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/audrism/globalopt}
  \item Report bugs at \url{https://github.com/audrism/globalopt/issues}
}

}
\author{
\strong{Maintainer}: Audris Mockus \email{audris@utk.edu}

Authors:
\itemize{
  \item Audris Mockus \email{audris@utk.edu}
  \item Jonas Mockus (author of the original MINIMUM/GlobalMinimum Fortran library (1989))
}

}
\keyword{internal}
